fix double

+++ app/src/main/java/com/example/kasapp/data/entity/DetailTransaksi.kt
@@ -58,6 +58,6 @@ data class DetailTransaksi(
     // BIARKAN DETAILNYA TETAP LENGKAP
     val namaMenuSaatTransaksi: String,
     val jumlah: Int,
-    val hargaSaatTransaksi: Double
+    val hargaSaatTransaksi: Int
 )
